requirements.txt
matplotlib>=3.4.2
numpy>=1.21.6
Pillow>=9.2.0
# scikit_image>=0.19.2
# skimage>=0.0
torch>=1.8.1+cu111
torchvision>=0.9.1+cu111
